requirements.txt
# 与 f2==0.0.1.7 声明一致；pydantic 放宽为 >=2.10 以便 Python 3.14 使用预编译 pydantic-core（避免本地编译 Rust/MSVC）。
# 安装完本文件后请执行: pip install "f2==0.0.1.7" --no-deps  （或运行 install-deps.bat）
aiofiles==24.1.0
aiosqlite==0.20.0
babel==2.13.0
black==24.10.0
browser-cookie3==0.20.1
click==8.1.7
cryptography==44.0.0
gmssl==3.2.2
httpx==0.27.2
importlib-resources==6.4.5
jsonpath-ng==1.6.1
m3u8==3.6.0
protobuf==5.28.3
pydantic>=2.10,<3
pyexecjs==1.5.1
pytest-asyncio==0.25.0
pytest==8.3.4
pyyaml==6.0.2
qrcode==8.0
rich==13.9.3
websockets-proxy==0.1.2
websockets<13.0
